examples/aflow_integration/requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.40.0
datasets>=2.14.0
accelerate>=0.20.0

# ROLL framework
# Install from: git+https://github.com/microsoft/ROLL.git

# AFlow framework
# Install from: git+https://github.com/geekan/AFlow.git

# Distributed training
ray[default]>=2.0.0
megatron-core>=0.5.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
jsonlines>=3.1.0

# Configuration
hydra-core>=1.3.0
omegaconf>=2.3.0

# Utilities
tqdm>=4.65.0
wandb>=0.15.0
tensorboard>=2.13.0

# Optional: Inference engines
# vllm>=0.2.0
# sglang>=0.1.0
